[Perf - part 2] Hoist `useEmail` hook (#8813)
