--- src/lib/hooks/useEmail.ts
@@ -1,35 +1,0 @@
-import {STALE} from '#/state/queries'
-import {useProfileQuery} from '#/state/queries/profile'
-import {useCheckEmailConfirmed} from '#/state/service-config'
-import {useSession} from '#/state/session'
-import {BSKY_SERVICE} from '../constants'
-import {getHostnameFromUrl} from '../strings/url-helpers'
-
-export function useEmail() {
-  const {currentAccount} = useSession()
-
-  const {data: profile} = useProfileQuery({
-    did: currentAccount?.did,
-    staleTime: STALE.INFINITY,
-  })
-
-  const checkEmailConfirmed = useCheckEmailConfirmed()
-
-  // Date set for 11 AM PST on the 18th of November
-  const isNewEnough =
-    !!profile?.createdAt &&
-    Date.parse(profile.createdAt) >= Date.parse('2024-11-18T19:00:00.000Z')
-
-  const isSelfHost =
-    currentAccount &&
-    getHostnameFromUrl(currentAccount.service) !==
-      getHostnameFromUrl(BSKY_SERVICE)
-
-  const needsEmailVerification =
-    !isSelfHost &&
-    checkEmailConfirmed &&
-    !currentAccount?.emailConfirmed &&
-    isNewEnough
-
-  return {needsEmailVerification}
-}
